add check for VSCode remote tunnel (#266)
In the VSCode Microsft distributed version you can run it "remotely". Like you start the IDE as a server-only at one PC and access it through Microsoft servers and its authentication on another PC, and inside the IDE you can access an integrated terminal that is running in the remote PC. And when I was using sudo for some commands the pam_usb checked for my puggled device and authorized it.

This is caused by VSCode not registering it's sesssion(s) in utmp.

Reference here: https://code.visualstudio.com/docs/remote/tunnels